# solution/main_server/swagger/swaggerDocumentationMainServer.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Football Analyzer API",
    "description": "API documentation for the Football application. Reminder: If you want to try a particular route which requires two or more parameters, and you want to filter by one or two of them, just insert an empty string ('''') in the field you want to leave empty.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "main express server"
    },
    {
      "url": "http://localhost:4000",
      "description": "second express server"
    },
    {
      "url": "http://localhost:8080",
      "description": "springboot server"
    }
  ],
  "paths": {
    "/clubs/main/get_clubs": {
      "get": {
        "summary": "Get clubs from the main server",
        "description": "Fetches the most popular clubs",
        "responses": {
          "200": {
            "description": "A list of clubs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MainClubs"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:3000"]
      }
    },
    "/clubs/main/get_filtered_clubs": {
      "get": {
        "summary": "Get filtered clubs from the main server",
        "description": "Fetches clubs by name, competition or country",
        "parameters": [
          {
            "in": "query",
            "name": "clubName",
            "schema": {
              "type": "string"
            },
            "description": "The name of the club"
          },
          {
            "in": "query",
            "name": "competition",
            "schema": {
              "type": "string"
            },
            "description": "The competition of the club"
          },
          {
            "in": "query",
            "name": "country",
            "schema": {
              "type": "string"
            },
            "description": "The country of the club"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Limit the number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of filtered clubs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MainFilteredClubs"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No clubs found matching the provided filters"
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:3000"]
      }
    },
    "/competitions/main/get_competitions": {
      "get": {
        "summary": "Get competitions from the main server",
        "description": "Fetches the most popular competitions",
        "responses": {
          "200": {
            "description": "A list of competitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MainCompetitions"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:3000"]
      }
    },
    "/competitions/main/get_filtered_competitions": {
      "get": {
        "summary": "Get filtered competitions from the main server",
        "description": "Fetches competitions by name,country or type",
        "parameters": [
          {
            "in": "query",
            "name": "competitionName",
            "schema": {
              "type": "string"
            },
            "description": "The name of the competition"
          },
          {
            "in": "query",
            "name": "country",
            "schema": {
              "type": "string"
            },
            "description": "The country of the competition"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "The type of the competition"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Limit the number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of filtered competitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MainFilteredCompetitions"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No competitions found matching the provided filters"
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:3000"]
      }
    },
    "/competitions/main/get_ranking": {
      "get": {
        "summary": "Get ranking from the main server",
        "description": "Fetches the raking of a specified domestic competition",
        "parameters": [
          {
            "in": "query",
            "name": "selectedCompetition",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ID of the competition"
          },
          {
            "in": "query",
            "name": "selectedYear",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The year of the competition"
          }
        ],
        "responses": {
          "200": {
            "description": "The raking of the competition specified",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MainRanking"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:3000"]
      }
    },
...
    "/games/express/get_games": {
      "get": {
        "summary": "Get games from the secondary express server",
        "description": "Fetches games by competition, year or round",
        "parameters": [
          {
            "in": "query",
            "name": "selectedCompetition",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ID of the competition"
          },
          {
            "in": "query",
            "name": "selectedYear",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The year of the competition"
          },
          {
            "in": "query",
            "name": "selectedRound",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The round of the competition"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Limit the number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of games",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpressGames"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No games found for this competition"
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:4000"]
      }
    },
    "/games/express/get_ranking": {
      "get": {
        "summary": "Get ranking from the secondary express server",
        "description": "Fetches the raking of a specified domestic competition",
        "parameters": [
          {
            "in": "query",
            "name": "selectedCompetition",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ID of the competition"
          },
          {
            "in": "query",
            "name": "selectedYear",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The year of the competition"
          }
        ],
        "responses": {
          "200": {
            "description": "The raking of the specified competition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpressRanking"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "tags": ["localhost:4000"]
      }
    },
...
  },
  "components": {
    "schemas": {
      "MainClubs": {
        "type": "object",
        "properties": {
          "club_id": {
            "type": "integer",
            "example": 31
          },
          "name": {
            "type": "string",
            "example": "Liverpool Football Club"
          },
          "domestic_competition_id": {
            "type": "string",
            "example": "GB1"
          },
          "domestic_competition_name": {
            "type": "string",
            "example": "premier-league"
          },
          "squad_size": {
            "type": "integer",
            "example": 26
          },
          "average_age": {
            "type": "number",
            "format": "float",
            "example": 26.2
          },
          "foreigners_number": {
            "type": "integer",
            "example": 21
          },
          "foreigners_percentage": {
            "type": "number",
            "format": "float",
            "example": 80.8
          },
          "national_team_players": {
            "type": "integer",
            "example": 17
          },
          "stadium_name": {
            "type": "string",
            "example": "Anfield"
          },
          "stadium_seats": {
            "type": "integer",
            "example": 54074
          },
          "net_transfer_record": {
            "type": "string",
            "example": "€-111.30m"
          },
          "last_season": {
            "type": "integer",
            "example": 2023
          }
        }
      },
      "MainFilteredClubs": {
        "type": "object",
        "properties": {
          "club_id": {
            "type": "integer",
            "example": 506
          },
          "name": {
            "type": "string",
            "example": "Juventus Football Club"
          },
          "domestic_competition_id": {
            "type": "string",
            "example": "IT1"
          },
          "domestic_competition_name": {
            "type": "string",
            "example": "serie-a"
          },
          "squad_size": {
            "type": "integer",
            "example": 26
          },
          "average_age": {
            "type": "number",
            "format": "float",
            "example": 26.7
          },
          "foreigners_number": {
            "type": "integer",
            "example": 14
          },
          "foreigners_percentage": {
            "type": "number",
            "format": "float",
            "example": 53.8
          },
          "national_team_players": {
            "type": "integer",
            "example": 16
          },
          "stadium_name": {
            "type": "string",
            "example": "Allianz Stadium"
          },
          "stadium_seats": {
            "type": "integer",
            "example": 41507
          },
          "net_transfer_record": {
            "type": "string",
            "example": "€-21.60m"
          },
          "last_season": {
            "type": "integer",
            "example": 2023
          }
        }
      },
      "MainCompetitions": {
        "type": "object",
        "properties": {
          "competition_id": {
            "type": "string",
            "example": "EL"
          },
          "competition_name": {
            "type": "string",
            "example": "europa-league"
          },
          "sub_type": {
            "type": "string",
            "example": "europa_league"
          },
          "type": {
            "type": "string",
            "example": "international_cup"
          },
          "country_id": {
            "type": "integer",
            "example": -1
          },
          "country_name": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "confederation": {
            "type": "string",
            "example": "europa"
          }
        }
      },
      "MainFilteredCompetitions": {
        "type": "object",
        "properties": {
          "competition_id": {
            "type": "string",
            "example": "EL"
          },
          "competition_name": {
            "type": "string",
            "example": "europa-league"
          },
          "sub_type": {
            "type": "string",
            "example": "europa_league"
          },
          "type": {
            "type": "string",
            "example": "international_cup"
          },
          "country_id": {
            "type": "integer",
            "example": -1
          },
          "country_name": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "confederation": {
            "type": "string",
            "example": "europa"
          }
        }
      },
      "MainRanking": {
        "type": "object",
        "properties": {
          "ClubId": {
            "type": "integer",
            "example": 46
          },
          "Club": {
            "type": "string",
            "example": "Football Club Internazionale Milano S.p.A."
          },
          "Points": {
            "type": "integer",
            "example": 32
          },
          "GamesPlayed": {
            "type": "integer",
            "example": 13
          },
          "GoalsScored": {
            "type": "integer",
            "example": 30
          },
          "GoalsConceded": {
            "type": "integer",
            "example": 7
          },
          "GoalsDifference": {
            "type": "integer",
            "example": 23
          }
        }
      },
...
      "ExpressGames": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65eb016dbd16e193a39faca2"
          },
          "game_id": {
            "type": "integer",
            "example": 2814762
          },
          "home_club_id": {
            "type": "integer",
            "example": 6195
          },
          "away_club_id": {
            "type": "integer",
            "example": 430
          },
          "home_club_goals": {
            "type": "integer",
            "example": 1
          },
          "away_club_goals": {
            "type": "string",
            "example": 0
          }
        }
      },
      "ExpressRanking": {
        "type": "object",
        "properties": {
          "ClubId": {
            "type": "integer",
            "example": 46
          },
          "Club": {
            "type": "string",
            "example": "Football Club Internazionale Milano S.p.A."
          },
          "Points": {
            "type": "integer",
            "example": 32
          },
          "GamesPlayed": {
            "type": "integer",
            "example": 13
          },
          "GoalsScored": {
            "type": "integer",
            "example": 30
          },
          "GoalsConceded": {
            "type": "integer",
            "example": 7
          },
          "GoalsDifference": {
            "type": "integer",
            "example": 23
          }
        }
      },
...
    }
  }
}